chore: Exporting CIDInfo class model.
