+++ lib/dart_cid.dart
@@ -1,3 +1,4 @@
 library dart_cid;
 
 export 'src/cid.dart';
+export 'src/models.dart';
